Minor drawing tweak for shelf-like containers
follow-up of https://github.com/twohoursonelife/OneLife/pull/42/commits/5062b1f0cbdb4b075144e45a458226ced34b477f, for objects with sprites that are rotated and with anchor not at the sprite center, such as yew bow. It still doesn't perfectly locate the lowest y among the sprites in an object. For example, when the Yew Bow object's sprite has a rotation of 0.625 it is drawn too high on a shelf. If applicable, consider the 0.999 trick in the commit mentioned above to skip the problematic sprite in containment position calculation.