Added containment transitions
Transitions can now be triggered when player interacts with a container. It can be specified to be triggered when (the first of / the last of / any of but not swapping / any of) (specific / any) object is being (put in/ taken out); and either / both of the container and the object can be transitioned.

For A + B = C + D, if A and C are the containers then this transition is triggered when things are taken out; if B and D are then otherwise. If the objects are left empty, then the transition is triggered when any object is taken out or put in. The "First / Last / Any except Swap / Any" part is specially flagged in the editor.

More details here: https://discord.com/channels/423293333864054833/944170642893316096/963865730758082641